Add files via upload
This is a running working copy of the program. i downloaded the current vm.c changed the name to Cases.c threw it on eustis. no compile time errors or warnings. i ran a test case provided by the teach and ignoring white space the only difference was that thier was no | to differentiate the stacks.
i changed some stuff around that were giving errors and coded it so  that when compiling on linux "./a.out mcode.pm0" uses the 2nd string.
so in the case i printed up it would look for "mcode.pm0". runs off "argv[1]" when looking for the file.
******only need to include the | between stacks and we are done*********